refactor(corpus): split index writers by responsibility (#0000)
